requirements.txt
streamlit
google-generativeai
sqlite3